[Bronze III] Title: 사탕 선생 고창영, Time: 4 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/2547. 사탕 선생 고창영/사탕 선생 고창영.cc
@@ -16,9 +16,9 @@ int dc[4] = { -1,1,0,0 };	// 좌우상하
 
 void solution()
 {
-	unsigned long long n;
+	int n;
 	cin >> n;
-	unsigned long long sum = 0;
+	int sum = 0;
 
 	forn(i, n)
 	{
